book-recommender/requirements.txt
pandas
numpy
matplotlib
seaborn
sentence-transformers
faiss-cpu
streamlit